chore(core): move saluki-event crate into saluki_core::data_model::event
