+++ lib/saluki-components/src/destinations/datadog/events/mod.rs
@@ -7,6 +7,7 @@ use http::{HeaderValue, Uri};
 use memory_accounting::{MemoryBounds, MemoryBoundsBuilder, UsageExpr};
 use saluki_common::task::HandleExt as _;
 use saluki_config::{GenericConfiguration, RefreshableConfiguration};
+use saluki_core::data_model::event::{eventd::EventD, DataType};
 use saluki_core::{
     components::{destinations::*, ComponentContext},
     observability::ComponentMetricsExt as _,
@@ -14,7 +15,6 @@ use saluki_core::{
     topology::interconnect::FixedSizeEventBuffer,
 };
 use saluki_error::{ErrorContext as _, GenericError};
-use saluki_event::{eventd::EventD, DataType};
 use saluki_io::{
     buf::{BytesBuffer, FrozenChunkedBytesBuffer},
     compression::CompressionScheme,
