+++ lib/saluki-components/src/sources/heartbeat/mod.rs
@@ -4,9 +4,9 @@ use async_trait::async_trait;
 use memory_accounting::{MemoryBounds, MemoryBoundsBuilder};
 use saluki_context::Context;
 use saluki_core::components::{sources::*, ComponentContext};
+use saluki_core::data_model::event::{metric::Metric, DataType, Event};
 use saluki_core::topology::OutputDefinition;
 use saluki_error::GenericError;
-use saluki_event::{metric::Metric, DataType, Event};
 use tokio::{select, time::interval};
 use tracing::{debug, error};
