+++ lib/saluki-components/src/transforms/preaggregation_filter/mod.rs
@@ -43,7 +43,7 @@ impl SynchronousTransform for PreaggregationFilter {
 #[cfg(test)]
 mod tests {
     use saluki_context::Context;
-    use saluki_event::{metric::Metric, Event};
+    use saluki_core::data_model::event::{metric::Metric, Event};
 
     use super::*;
