+++ lib/saluki-core/Cargo.toml
@@ -7,27 +7,32 @@ repository = { workspace = true }
 
 [dependencies]
 async-trait = { workspace = true }
+bitmask-enum = { workspace = true }
+ddsketch-agent = { workspace = true }
 futures = { workspace = true }
 indexmap = { workspace = true, features = ["std"] }
 memory-accounting = { workspace = true }
 metrics = { workspace = true }
 metrics-util = { workspace = true, features = ["handles", "recency", "registry"] }
+ordered-float = { workspace = true }
 paste = { workspace = true }
 pin-project = { workspace = true }
 saluki-common = { workspace = true }
 saluki-context = { workspace = true }
 saluki-error = { workspace = true }
-saluki-event = { workspace = true }
 saluki-health = { workspace = true }
 saluki-metrics = { workspace = true }
+serde = { workspace = true, features = ["derive"] }
 slab = { workspace = true }
 smallvec = { workspace = true }
 snafu = { workspace = true }
+stringtheory = { workspace = true }
 tokio = { workspace = true, features = ["macros", "rt", "rt-multi-thread", "sync", "time"] }
 tokio-util = { workspace = true, features = ["time"] }
 tracing = { workspace = true, features = ["std"] }
 
 [dev-dependencies]
+proptest = { workspace = true }
 similar-asserts = { workspace = true, features = ["unicode"] }
 tokio-test = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["fmt"] }
