+++ lib/saluki-core/src/components/destinations/builder.rs
@@ -1,10 +1,9 @@
 use async_trait::async_trait;
 use memory_accounting::MemoryBounds;
 use saluki_error::GenericError;
-use saluki_event::DataType;
 
 use super::Destination;
-use crate::components::ComponentContext;
+use crate::{components::ComponentContext, data_model::event::DataType};
 
 /// A destination builder.
 ///
