+++ lib/saluki-core/src/data_model/event/metric/value/scalar.rs
