+++ lib/saluki-core/src/data_model/mod.rs
@@ -1,0 +1,2 @@
+//! Data model for Saluki components.
+pub mod event;
