+++ lib/saluki-core/src/lib.rs
@@ -9,6 +9,7 @@ pub mod reexport {
 
 pub mod components;
 pub mod constants;
+pub mod data_model;
 pub mod observability;
 pub mod pooling;
 pub mod state;
