+++ lib/saluki-core/src/topology/interconnect/event_buffer.rs
@@ -1,8 +1,9 @@
 use std::{collections::VecDeque, fmt};
 
-use saluki_event::{DataType, Event};
-
-use crate::pooling::{helpers::pooled_newtype, Clearable, ObjectPool};
+use crate::{
+    data_model::event::{DataType, Event},
+    pooling::{helpers::pooled_newtype, Clearable, ObjectPool},
+};
 
 /// A double-ended queue implemented with a ring buffer that has a fixed capacity at creation.
 struct FixedSizeVecDeque<T>(VecDeque<T>);
@@ -218,15 +219,16 @@ where
 mod tests {
     use std::collections::VecDeque;
 
-    use saluki_event::{
-        eventd::EventD,
-        metric::Metric,
-        service_check::{CheckStatus, ServiceCheck},
-        DataType, Event,
-    };
-
     use super::{FixedSizeEventBuffer, FixedSizeEventBufferInner};
-    use crate::pooling::{helpers::get_pooled_object_via_builder, Clearable as _};
+    use crate::{
+        data_model::event::{
+            eventd::EventD,
+            metric::Metric,
+            service_check::{CheckStatus, ServiceCheck},
+            DataType, Event,
+        },
+        pooling::{helpers::get_pooled_object_via_builder, Clearable as _},
+    };
 
     #[test]
     fn capacity() {
