+++ lib/saluki-core/src/topology/test_util.rs
@@ -1,14 +1,16 @@
 use async_trait::async_trait;
 use memory_accounting::{MemoryBounds, MemoryBoundsBuilder};
 use saluki_error::GenericError;
-use saluki_event::DataType;
 
 use super::OutputDefinition;
-use crate::components::{
-    destinations::{Destination, DestinationBuilder, DestinationContext},
-    sources::{Source, SourceBuilder, SourceContext},
-    transforms::{Transform, TransformBuilder, TransformContext},
-    ComponentContext,
+use crate::{
+    components::{
+        destinations::{Destination, DestinationBuilder, DestinationContext},
+        sources::{Source, SourceBuilder, SourceContext},
+        transforms::{Transform, TransformBuilder, TransformContext},
+        ComponentContext,
+    },
+    data_model::event::DataType,
 };
 
 struct TestSource;
