--- lib/saluki-event/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "saluki-event"
-version = "0.1.0"
-edition = { workspace = true }
-license = { workspace = true }
-repository = { workspace = true }
-
-[dependencies]
-bitmask-enum = { workspace = true }
-ddsketch-agent = { workspace = true }
-ordered-float = { workspace = true }
-saluki-context = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-smallvec = { workspace = true, features = ["const_generics", "drain_filter", "union"] }
-stringtheory = { workspace = true }
-
-[dev-dependencies]
-proptest = { workspace = true }
